fix cam

+++ middleware/errorHandler.js
@@ -1,8 +1,9 @@
 const deleteFile = require('../helpers/deleteFileGcs');
 
 function errorHandler(err, req, res, next) {
-    console.log(err, `<<<<<<<<<<<,,,,,,,,,,,,,,,,`);
-    if (req.file) {
+    console.log(err.message, `<<<<<<<<<<<,,,,,,,,,,,,,,,,`);
+    console.log(err.stack, `<<<<<<<<<<<,,,,,,,,,,,,,,,,`);
+    if (req.file && req.file.cloudStoragePublicUrl) {
         let file = req.file.cloudStoragePublicUrl.split('/');
         let fileName = file[file.length - 1];
         deleteFile(fileName)
